refactor: extract option-selection orchestration into DepositOptionCoordinator (#4991)
Move the per-option selection orchestration (selectDepositOption dispatch,
handleWithdrawSecuredAsset, requireSecuredAssetInboundAddress,
handleRemoveCacaoOption, resetTextFields) and the switchInboundJob /
withdrawSecuredAssetJob lifecycle out of DepositFormViewModel into a new
assisted-injected DepositOptionCoordinator. The VM keeps viewModelScope
ownership and forwards selectDepositOption as a thin public call; the
coordinator drives the existing loader seams and writes state back via the
shared StateFlow / form fields. Pure move, no behavioral change.
